Added support for template service units
Allow pulling instances of templated service units into the initramfs.
For example if a service `my-fancy-tool@my-arg.service` is wanted or
required by a target inside the initrd, the service file
`my-fancy-tool@.service` and the symlink
`<target>.wants/my-fancy-tool@my-arg.service` will now be added to the
initrd image file.