Added Support for Preserving Additional Users/Groups
do_secret_clean() in initrd-build.sh wipes all user accounts besides
a hard coded list. If a user would like to add a service that needs
to run as a specific user, they previously could not because the
user would be removed.

This patch adds a config option to preserve additional accounts and
reworks the build script a little to make this it cleaner.